+++ crates/holochain/src/core/ribosome/guest_callback/entry_defs.rs
@@ -182,7 +182,6 @@ mod slow_tests {
     use crate::fixt::RealRibosomeFixturator;
     use crate::fixt::ZomeCallHostAccessFixturator;
     use ::fixt::prelude::*;
-    use holochain_state::host_fn_workspace::HostFnWorkspace;
     use holochain_types::prelude::*;
     use holochain_wasm_test_utils::TestWasm;
     pub use holochain_zome_types::entry_def::EntryVisibility;
@@ -205,19 +204,7 @@ mod slow_tests {
 
     #[tokio::test(flavor = "multi_thread")]
     async fn test_entry_defs_index_lookup() {
-        let test_env = holochain_state::test_utils::test_cell_env();
-        let test_cache = holochain_state::test_utils::test_cache_env();
-        let env = test_env.env();
-
-        let author = fake_agent_pubkey_1();
-        crate::test_utils::fake_genesis(env.clone()).await.unwrap();
-
-        let workspace = HostFnWorkspace::new(env.clone(), test_cache.env(), author)
-            .await
-            .unwrap();
-
-        let mut host_access = fixt!(ZomeCallHostAccess);
-        host_access.workspace = workspace;
+        let host_access = fixt!(ZomeCallHostAccess, Predictable);
         let output: () =
             crate::call_test_ribosome!(host_access, TestWasm::EntryDefs, "assert_indexes", ())
                 .unwrap();